refactor(transport): split vsock.rs into submodules
- vsock/addr.rs: VsockAddr and DEFAULT_AGENT_PORT
- vsock/stream.rs: VsockStream with AsyncRead/AsyncWrite impl
- vsock/transport.rs: VsockTransport, VsockSender, VsockReceiver, read_framed
- vsock/listener.rs: VsockListener and TransportListener impl
- vsock/linux.rs: Linux AF_VSOCK platform code
- vsock/darwin.rs: macOS Virtualization.framework platform code
- vsock/mod.rs: module declarations, re-exports, tests
